[FIX] mail: show display_name for contact without name and email
Steps to reproduce
===============
1. Create a contact of type invoice address without name and email.
2. Go to any app with chatter.
3. Add this user to the recipient

----> Tag with label `Unnamed` will be shown.

Along with this, if there is any follower without a name and email,
`false` was shown in follower list and in the dialog to edit the
subscription.

After this commit, we will use the display_name as a fallback to show in
those places.

Note: Added/adapted tests to ensure the correct behaviour. Also some
mocked models were missing the definition for computing correct display
name. Thus, compute methods are overriden in those mocked models.

Task-4812554

X-original-commit: acb5f6ff3ba0b4c397b949beef6330d25a18d0a5