[IMP] html_builder*: use event listener to limit selection in container *: website
Before this commit: we can select the whole editing area in website
builder by mouse selection or ctrl+a

After this commit: we introduced a new plugin in the html_builder to
restrict the selections to the closest div element or p element. We also
moved the click listener in builder_options_plugin.js to properly update
the blue container to properly align with the restricted selection.

We have the restriction function called on event keydown, mouseup and
touchend. The listeners are set this way to avoid unnecessary selection
restriction during the selection making. The keydown event is to handle
ctrl+a to select the whole block and then restrict it. The mouseup and
touchend handle the cases where we make the selection by mouse or touch.

We introduced two concepts special blocks (with text in non div) and
uncrossable. In general cases, we restrict the selection to the closes
div element. However, for special blocks, we restrict selection in
closest p element. uncrossable can be defined by tag name or classes.
They are the elements that the selection should not cross, these
elements can appear in a div element and the restricted selection should
stop before crossing them.

When we do select all, we first check if we should select the whole div
element or the p element, then we extend the selection to one direction
+ correct the selection on uncrossable, then we do the same for the
other direction.

When we make selection by mouse, we first check if the current selection
is beyond the div or p element it should be and we set the selection
inside if it is. Then we check inside the div or p element, if there are
uncrossable, we correct the selection by stopping the selection ahead of
them.

task-4901968