[IMP] website, *: dynamic snippet products with title on left
*=website_sale

Before this commit, when the section title is placed on the
left and the content-width is not max in total four elements are
displayed (either two or one in each row).

This commit ensures that two items are displayed side-by-side in
"s_dynamic_snippet_products" when the section title is placed on the
left and the content-width is not max (container-fluid).

task-5090468