[FIX] purchase_stock: fix empty picking creation by po.
This commit fixes the problem of creating an empty picking when
confirming a PO with negative quantity for the product.

Before this commit:
When creating a PO for a vendor with a negative quantity for a product,
it creates an OUT picking with that same quantity to the vendor as
a return which is correct. However, it also creates an empty IN picking
which is not needed.

After this commit:
Only the OUT picking is created for the order lines with negative
quantity as a return and if there are order lines with positive quantity,
it creates a normal IN picking as expected.

Also as a side improvement, Now all updates to purchase order lines are
reflected correctly to the existing po's pickings.

Enterprise PR: https://github.com/odoo/enterprise/pull/108890

Task-5074899